requirements.txt
google-api-python-client==2.110.0
google-auth-oauthlib==1.1.0
google-auth==2.23.4
qdrant-client==1.7.0
openai==1.3.8
langchain==0.0.348
langchain-text-splitters==0.0.1
python-dotenv==1.0.0
python-docx==1.1.0
pypdf==3.17.4
# Image processing libraries
Pillow==10.1.0
pytesseract==0.3.10
requests==2.31.0
# HTML/Markup processing libraries
beautifulsoup4==4.12.2
lxml==4.9.3